backend/requirements.txt
# Core Framework
fastapi==0.115.9
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
sqlmodel==0.0.24
alembic==1.16.4
psycopg2-binary==2.9.10

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.11.4
pydantic-core==2.33.2

# Environment & Configuration
python-dotenv==1.1.0

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1

# Development
black==25.1.0
flake8==7.3.0
mypy==1.17.1

# Production
gunicorn==23.0.0

# Additional dependencies (exact versions from current installation)
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
coverage==7.10.4
cryptography==44.0.3
ecdsa==0.19.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.45.3
typing_extensions==4.13.2
watchfiles==1.0.5
websockets==15.0.1
